Revert deps between lila and playJson
`playJson` only provides extensions and generic type class
transformation. A new file `lila.Json` is created to provide
lila's specific instance for play-json type class.
