+++ build.sbt
@@ -39,6 +39,16 @@ lazy val model: Project = Project("model", file("model"))
     commonSettings,
     name := "scalalib-model"
   )
+  .dependsOn(core, playJson)
+
+lazy val playJson: Project = Project("playJson", file("playJson"))
+  .settings(
+    commonSettings,
+    name := "scalalib-play-json",
+    libraryDependencies ++= List(
+      "org.playframework" %% "play-json" % "3.0.2"
+    )
+  )
   .dependsOn(core)
 
 // functions that useful for lila & lila-ws
@@ -52,17 +62,7 @@ lazy val lila: Project = Project("lila", file("lila"))
       "org.scalameta"                %% "munit"     % "1.0.0-M11" % Test
     )
   )
-  .dependsOn(core, model)
-
-lazy val playJson: Project = Project("playJson", file("playJson"))
-  .settings(
-    commonSettings,
-    name := "scalalib-play-json",
-    libraryDependencies ++= List(
-      "org.playframework" %% "play-json" % "3.0.2"
-    )
-  )
-  .dependsOn(core, lila)
+  .dependsOn(core, model, playJson)
 
 lazy val root = project
   .in(file("."))
