+++ playJson/src/main/scala/Json.scala
@@ -7,7 +7,6 @@ import scala.util.NotGiven
 import alleycats.Zero
 
 import scalalib.newtypes.SameRuntime
-import scalalib.model.*
 import scalalib.time.toMillis
 
 object Json:
@@ -80,51 +79,3 @@ object Json:
   )
 
   given Writes[Instant] = writeAs(_.toMillis)
-
-  // given Writes[chess.Color] = writeAs(_.name)
-  //
-  // given Reads[Uci] = Reads
-  //   .of[String]
-  //   .flatMapResult: str =>
-  //     JsResult.fromTry(Uci(str).toTry(s"Invalid UCI: $str"))
-  // given Writes[Uci] = writeAs(_.uci)
-  //
-  // given Reads[LilaOpeningFamily] = Reads[LilaOpeningFamily]: f =>
-  //   f.get[String]("key")
-  //     .flatMap(LilaOpeningFamily.find)
-  //     .fold[JsResult[LilaOpeningFamily]](JsError(Nil))(JsSuccess(_))
-  //
-  // given NoJsonHandler[chess.Square] with {}
-  //
-  // given OWrites[Crazyhouse.Pocket] = OWrites: p =>
-  //   JsObject:
-  //     p.flatMap((role, nb) => Option.when(nb > 0)(role.name -> JsNumber(nb)))
-  //
-  // given OWrites[chess.variant.Crazyhouse.Data] = OWrites: v =>
-  //   PlayJson.obj("pockets" -> v.pockets.all)
-  //
-
-  given Writes[MaxPerPage] with
-    def writes(m: MaxPerPage) = JsNumber(m.value)
-
-  import scalalib.paginator.Paginator
-  given paginatorWrite[A: Writes]: OWrites[Paginator[A]] = OWrites[Paginator[A]]: p =>
-    PlayJson.obj(
-      "currentPage"        -> p.currentPage,
-      "maxPerPage"         -> p.maxPerPage,
-      "currentPageResults" -> p.currentPageResults,
-      "nbResults"          -> p.nbResults,
-      "previousPage"       -> p.previousPage,
-      "nextPage"           -> p.nextPage,
-      "nbPages"            -> p.nbPages
-    )
-
-  // import lila.core.LightUser
-  // given lightUserWrites: OWrites[LightUser] = OWrites(lightUser.write)
-  // object lightUser:
-  //   def write(u: LightUser): JsObject = writeNoId(u) + ("id" -> JsString(u.id.value))
-  //   def writeNoId(u: LightUser): JsObject = PlayJson
-  //     .obj("name" -> u.name)
-  //     .add("title", u.title)
-  //     .add("flair", u.flair)
-  //     .add("patron", u.isPatron)
